requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
attrs==19.3.0             # via jsonschema, pytest
autograd-gamma==0.4.2     # via lifelines
autograd==1.3             # via autograd-gamma, lifelines
backcall==0.1.0           # via ipython
bash-kernel==0.7.2
beautifulsoup4==4.9.1
bleach==3.1.4             # via nbconvert
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
coverage==4.5.4           # via nbval
cycler==0.10.0            # via matplotlib
decorator==4.4.1          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
future==0.18.2            # via autograd
idna==2.10                # via requests
ipykernel==5.1.4          # via ipywidgets, jupyter, jupyter-console, nbval, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.12.0           # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1
jedi==0.16.0              # via ipython
jinja2==2.11.1            # via jupyterlab, jupyterlab-server, nbconvert, notebook
json5==0.9.0              # via jupyterlab-server
jsonschema==3.2.0         # via jupyterlab-server, nbformat
jupyter-client==5.3.4     # via ipykernel, jupyter-console, nbval, notebook, qtconsole
jupyter-console==6.1.0    # via jupyter
jupyter-core==4.6.1       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
jupyterlab-server==1.0.6  # via jupyterlab
jupyterlab==1.2.6
jupytext==1.3.3
kiwisolver==1.1.0         # via matplotlib
lifelines==0.25.1
markupsafe==1.1.1         # via jinja2
matplotlib==3.1.3
mistune==0.8.4            # via nbconvert
more-itertools==8.2.0     # via pytest
nbconvert==5.6.1          # via jupyter, notebook
nbformat==5.0.4           # via ipywidgets, jupytext, nbconvert, nbval, notebook
nbval==0.9.4
notebook==6.0.3           # via jupyter, jupyterlab, jupyterlab-server, widgetsnbextension
numpy==1.18.1
packaging==20.1           # via pytest
pandas==1.0.1
pandocfilters==1.4.2      # via nbconvert
parso==0.6.1              # via jedi
patsy==0.5.1              # via lifelines, statsmodels
pexpect==4.8.0            # via bash-kernel, ipython
pickleshare==0.7.5        # via ipython
pip-tools==4.4.1
plotly==4.5.0
pluggy==0.13.1            # via pytest
prometheus-client==0.7.1  # via notebook
prompt-toolkit==3.0.3     # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
py==1.8.1                 # via pytest
pygments==2.5.2           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.4.6          # via matplotlib, packaging
pyrsistent==0.15.7        # via jsonschema
pytest==5.3.5             # via nbval
python-dateutil==2.8.1    # via jupyter-client, matplotlib, pandas
pytz==2019.3              # via pandas
pyyaml==5.3               # via jupytext
pyzmq==18.1.1             # via jupyter-client, notebook
qtconsole==4.6.0          # via jupyter
requests==2.24.0
retrying==1.3.3           # via plotly
scipy==1.4.1              # via autograd-gamma, lifelines, seaborn, statsmodels
seaborn==0.10.0
send2trash==1.5.0         # via notebook
six==1.14.0               # via bleach, cycler, jsonschema, nbval, packaging, patsy, pip-tools, plotly, pyrsistent, python-dateutil, retrying, traitlets
soupsieve==2.0.1          # via beautifulsoup4
statsmodels==0.11.0
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
tornado==6.0.3            # via ipykernel, jupyter-client, jupyterlab, notebook, terminado
tqdm==4.42.1
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
urllib3==1.25.10          # via requests
wcwidth==0.1.8            # via prompt-toolkit, pytest
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets
xlrd==1.2.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
